Decouple KBookmarksMenu from KActionCollection
Summary: The usage of KActionCollection makes KBookmarks depend on KXMLGui. To allow to get rid of this rather heavy dependency for KF6 a new constructor is introduced that does not take a KActionCollection. Instead apps that use KBookmarksMenu and KXMLGui are encouraged to manually add the relevant actions to their action collection. It's not pretty, but better than depending on KXMLGui, especially given that KIO depends on KBookmarks and thus on KXMLGui.

Test Plan: Tested with patched Dolphin (new constructor is used) and unpatched Konsole.

Reviewers: #frameworks, dfaure

Reviewed By: dfaure

Subscribers: kossebau, dfaure, apol, kde-frameworks-devel

Tags: #frameworks

Maniphest Tasks: T12056

Differential Revision: https://phabricator.kde.org/D25660