Use KDirWatch::self method instead of custom instance
In case of KRunner, we'd create manager in a different thread, but
delete it in the main thread. This means that our main thread cleans up
the objects using qAddPostRoutine, but QThreadStorage has already
deleted the KDirWatchPrivate entries.

This doesn't fix all thread safety issues, but the most critical
problem.